Move `simple_triangulate` to own submodule
