+++ nanomesh/mesh2d/__init__.py
@@ -1,4 +1,4 @@
-from .helpers import compare_mesh_with_image, pad, simple_triangulate
+from .helpers import compare_mesh_with_image, pad
 from .mesher import Mesher2D, generate_2d_mesh
 
 __all__ = [
@@ -6,5 +6,4 @@ __all__ = [
     'generate_2d_mesh',
     'Mesher2D',
     'pad',
-    'simple_triangulate',
 ]
