+++ nanomesh/mesh2d/helpers.py
@@ -1,17 +1,15 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Dict, List, Tuple
+from typing import TYPE_CHECKING
 
 import matplotlib.pyplot as plt
 import numpy as np
-import triangle as tr
 from scipy.spatial.distance import cdist
 
 from ..region_markers import RegionMarker
 
 if TYPE_CHECKING:
     from nanomesh.mesh import LineMesh, TriangleMesh
-    from nanomesh.mesh_container import MeshContainer
 
 
 def compare_mesh_with_image(image: np.ndarray, mesh: TriangleMesh):
@@ -38,52 +36,6 @@ def compare_mesh_with_image(image: np.ndarray, mesh: TriangleMesh):
     ax.set_yticks([])
 
     return ax
-
-
-def simple_triangulate(points: np.ndarray,
-                       *,
-                       segments: np.ndarray = None,
-                       regions: List[Tuple[float, ...]] = None,
-                       opts: str = '') -> MeshContainer:
-    """Simple triangulation using `triangle`.
-
-    Parameters
-    ----------
-    points : i,2 np.ndarray
-        Vertex coordinates.
-    segments : j,2 np.ndarray, optional
-        Index array describing segments.
-        Segments are edges whose presence in the triangulation
-        is enforced (although each segment may be subdivided into smaller
-        edges). Each segment is specified by listing the indices of its
-        two endpoints. A closed set of segments describes a contour.
-    regions : k,2 np.ndarray, optional
-        Coordinates describing regions. A region is a coordinate inside
-        (e.g. at the center) of a region/contour (i.e. enclosed by segments).
-    opts : str, optional
-        Additional options passed to `triangle.triangulate` documented here:
-        https://rufat.be/triangle/API.html#triangle.triangulate
-
-    Returns
-    -------
-    mesh : MeshContainer
-        Triangle mesh
-    """
-    from nanomesh.mesh_container import MeshContainer
-
-    triangle_dict_in: Dict['str', Any] = {'vertices': points}
-
-    if segments is not None:
-        triangle_dict_in['segments'] = segments
-
-    if regions is not None:
-        triangle_dict_in['regions'] = regions
-
-    triangle_dict_out = tr.triangulate(triangle_dict_in, opts=opts)
-
-    mesh = MeshContainer.from_triangle_dict(triangle_dict_out)
-
-    return mesh
 
 
 def pad(
