+++ nanomesh/mesh3d/mesher.py
@@ -9,10 +9,10 @@ import numpy as np
 from skimage import measure, morphology
 
 from nanomesh._mesh_shared import BaseMesher
-from nanomesh.mesh2d import simple_triangulate
 from nanomesh.volume import Volume
 
 from ..region_markers import RegionMarker, RegionMarkerLike
+from ..triangulate import simple_triangulate
 from .bounding_box import BoundingBox
 from .helpers import pad
